requirements.txt
altgraph==0.10.2
backports-abc==0.4
bdist-mpkg==0.5.0
BeautifulSoup==3.2.1
beautifulsoup4==4.5.1
bonjour-py==0.3
bs4==0.0.1
cement==2.10.2
certifi==2016.9.26
click==6.7
Django==1.10.2
enum34==1.1.6
Flask==0.12
fusepy==2.0.2
futures==3.2.0
grpc==0.3.post19
grpcio==1.7.0
grpcio-tools==1.7.0
inflection==0.3.1
itsdangerous==0.24
Jinja2==2.9.4
macholib==1.5.1
MarkupSafe==0.23
matplotlib==1.3.1
modulegraph==0.10.4
mpld3==0.3
msgpack-python==0.4.8
nose==1.3.7
numpy==1.8.0rc1
peewee==2.8.5
protobuf==3.5.0.post1
py2app==0.7.3
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==0.13.1
pyparsing==2.0.1
python-dateutil==1.5
pytz==2013.7
requests==2.11.1
scipy==0.13.0b1
semantic-version==2.6.0
simplejson==3.8.2
singledispatch==3.4.0.3
six==1.11.0
tornado==4.4.2
virtualenv==15.1.0
Werkzeug==0.11.15
xattr==0.6.4
zope.interface==4.1.1
